Define início do app pelo Gallery.js
