+++ dojos/20170927-react-native/base/App.js
@@ -1,30 +1,9 @@
-import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
+import { StackNavigator } from 'react-navigation';
 
-export default class App extends React.Component {
-  render() {
-    return (
-      <View style={styles.container}>
-        <Text style={styles.title}>Dojo #8</Text>
-        <Text>Campinas Front-End</Text>
-        <Text>Let's start by editing <Text style={styles.bold}>/Dojo/App.js</Text></Text>
-      </View>
-    );
-  }
-}
+import Gallery from './screens/Gallery';
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  title: {
-    fontSize: 22,
-    fontWeight: 'bold',
-  },
-  bold: {
-    fontWeight: 'bold',
-  }
+const SimpleApp = StackNavigator({
+  Home: { screen: Gallery },
 });
+
+export default SimpleApp;
