+++ dojos/20170927-react-native/base/screens/Gallery.js
@@ -1,0 +1,32 @@
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+
+class Gallery extends React.Component {
+  render() {
+    return (
+      <View style={styles.container}>
+        <Text style={styles.title}>Dojo #8</Text>
+        <Text>Campinas Front-End</Text>
+        <Text>Vamos começar editando <Text style={styles.bold}>/screens/Gallery.js</Text></Text>
+      </View>
+    );
+  }
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  title: {
+    fontSize: 22,
+    fontWeight: 'bold',
+  },
+  bold: {
+    fontWeight: 'bold',
+  }
+});
+
+export default Gallery;
